+++ extension/src/Settings/SettingsMap.ts
@@ -64,4 +64,7 @@ export const settingsMap = new Map<string, keyof Settings>([
   ["NAB.EnableTroubleshootingCommands", "enableTroubleshootingCommands"],
   ["NAB.EnableXliffCache", "enableXliffCache"],
   ["NAB.PreferLockedTranslations", "preferLockedTranslations"],
+
+  // Other extension's settings:
+  ["al.packageCachePath", "packageCachePath"],
 ]);